refactor: 새로 정의한 타입으로 import / DP-318
