+++ src/features/map/components/MapPanelItem/MapPanelProjectItem/MapPanelProjectItem.tsx
@@ -1,4 +1,4 @@
-import type { ProjectPanelItem } from '@/features/map/types';
+import type { ProjectPanelItem } from '@/features/map/schemas/mapItemSchema';
 import { MAP_ITEM_STATUS_LABELS, TeamStatus } from '@/features/map/types/common';
 
 import styles from '../MapPanelItem.module.scss';
