+++ src/features/map/components/MapPanelItem/MapPanelStudyItem/MapPanelStudyItem.tsx
@@ -1,4 +1,4 @@
-import type { StudyPanelItem } from '@/features/map/types';
+import type { StudyPanelItem } from '@/features/map/schemas/mapItemSchema';
 import { MAP_ITEM_STATUS_LABELS, TeamStatus } from '@/features/map/types/common';
 
 import styles from '../MapPanelItem.module.scss';
